+++ core/util/fsutils.go
@@ -22,6 +22,7 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 )
 
@@ -204,4 +205,10 @@ func CreateFile(dir string, name string) (string, error) {
 	}
 
 	return src, nil
+}
+
+// FileExtension extract file extension from file name
+func FileExtension(file string) string {
+	ext := filepath.Ext(file)
+	return strings.TrimSpace(ext)
 }