+++ modules/crawler/pipe/url_filter_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package pipe
 
 import (
+	"github.com/infinitbyte/gopa/core/config"
 	"github.com/infinitbyte/gopa/core/model"
 	"github.com/infinitbyte/gopa/core/pipeline"
 	"github.com/stretchr/testify/assert"
@@ -35,11 +36,16 @@ func TestUrlFilter(t *testing.T) {
 	task.Breadth = 1
 	task.Host = "elasticsearch.cn"
 
+	snapshot := model.Snapshot{}
+	snapshot.Path = "/"
+	snapshot.File = "default.html"
+
 	context.Set(CONTEXT_CRAWLER_TASK, &task)
+	context.Set(CONTEXT_CRAWLER_SNAPSHOT, &snapshot)
 
 	parse := UrlExtFilterJoint{}
 	parse.Process(context)
-	assert.Equal(t, false, context.IsBreak())
+	assert.Equal(t, false, context.IsEnd())
 
 	task = model.Task{}
 	task.Url = "mailto:g"
@@ -48,7 +54,7 @@ func TestUrlFilter(t *testing.T) {
 
 	context.Set(CONTEXT_CRAWLER_TASK, &task)
 	parse.Process(context)
-	assert.Equal(t, true, context.IsErrorExit())
+	assert.Equal(t, true, context.IsExit())
 
 	task = model.Task{}
 	task.Url = "asfasdf.gif"
@@ -57,6 +63,78 @@ func TestUrlFilter(t *testing.T) {
 
 	context.Set(CONTEXT_CRAWLER_TASK, &task)
 	parse.Process(context)
-	assert.Equal(t, true, context.IsErrorExit())
+	assert.Equal(t, true, context.IsExit())
+
+}
+
+func TestRuleCheck(t *testing.T) {
+	rule := config.Rules{}
+	rule.MustNot = &config.Rule{}
+	rule.MustNot.Prefix = append(rule.MustNot.Prefix, "prefix")
+	rule.MustNot.Contain = append(rule.MustNot.Contain, "contain")
+	rule.MustNot.Suffix = append(rule.MustNot.Suffix, "suffix")
+
+	assert.Equal(t, false, checkRule(rule, "prefix.google.com"))
+	assert.Equal(t, false, checkRule(rule, "www.contain.com"))
+	assert.Equal(t, false, checkRule(rule, "www.google.suffix"))
+
+	rule = config.Rules{}
+	rule.Must = &config.Rule{}
+	rule.Must.Prefix = append(rule.Must.Prefix, "prefix")
+	rule.Must.Contain = append(rule.Must.Contain, "contain")
+	rule.Must.Suffix = append(rule.Must.Suffix, "suffix")
+
+	assert.Equal(t, false, checkRule(rule, "prefix.google.com"))
+	assert.Equal(t, false, checkRule(rule, "www.contain.com"))
+	assert.Equal(t, false, checkRule(rule, "www.google.suffix"))
+	assert.Equal(t, true, checkRule(rule, "prefix.contain.suffix"))
+
+	rule = config.Rules{}
+	rule.Must = &config.Rule{}
+	rule.MustNot = &config.Rule{}
+
+	rule.Should = &config.Rule{}
+	rule.Should.Prefix = append(rule.Should.Prefix, "prefix")
+	rule.Should.Contain = append(rule.Should.Contain, "contain")
+	rule.Should.Suffix = append(rule.Should.Suffix, "suffix")
+
+	assert.Equal(t, true, checkRule(rule, "prefix.google.com"))
+	assert.Equal(t, true, checkRule(rule, "www.contain.com"))
+	assert.Equal(t, true, checkRule(rule, "www.google.suffix"))
+	assert.Equal(t, true, checkRule(rule, "prefix.contain.suffix"))
+	assert.Equal(t, false, checkRule(rule, "www.baidu.com"))
+
+	rule = config.Rules{}
+	rule.Must = &config.Rule{}
+	rule.MustNot = &config.Rule{}
+	rule.MustNot.Prefix = append(rule.MustNot.Prefix, "non-exists")
+
+	rule.Should = &config.Rule{}
+	rule.Should.Prefix = append(rule.Should.Prefix, "prefix")
+	rule.Should.Contain = append(rule.Should.Contain, "contain")
+	rule.Should.Suffix = append(rule.Should.Suffix, "suffix")
+
+	assert.Equal(t, true, checkRule(rule, "prefix.google.com"))
+	assert.Equal(t, true, checkRule(rule, "www.contain.com"))
+	assert.Equal(t, true, checkRule(rule, "www.google.suffix"))
+	assert.Equal(t, true, checkRule(rule, "prefix.contain.suffix"))
+	assert.Equal(t, false, checkRule(rule, "www.baidu.com"))
+
+	rule = config.Rules{}
+	rule.Must = &config.Rule{}
+	rule.Must.Contain = append(rule.Must.Contain, ".")
+	rule.MustNot = &config.Rule{}
+	rule.MustNot.Prefix = append(rule.MustNot.Prefix, "non-exists")
+
+	rule.Should = &config.Rule{}
+	rule.Should.Prefix = append(rule.Should.Prefix, "prefix")
+	rule.Should.Contain = append(rule.Should.Contain, "contain")
+	rule.Should.Suffix = append(rule.Should.Suffix, "suffix")
+
+	assert.Equal(t, true, checkRule(rule, "prefix.google.com"))
+	assert.Equal(t, true, checkRule(rule, "www.contain.com"))
+	assert.Equal(t, true, checkRule(rule, "www.google.suffix"))
+	assert.Equal(t, true, checkRule(rule, "prefix.contain.suffix"))
+	assert.Equal(t, false, checkRule(rule, "www.baidu.com"))
 
 }